Adding experimental support for the "Cotira" CMake module.
This module adds support for PCH's and unity builds.  After a lot
of hacking to fix problems, the PCH mechanism was seen to be working,
and precompiled all headers outside of the src folder.  However,
it didn't seem to lead to any noticeable speedup, so this is being
set aside for now.